pkg/plugin-request/requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.103.0
h11==0.14.0
idna==3.4
pydantic==1.10.9
PyYAML==6.0.1
requests==2.30.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0
